Norm: norm texture and texture manager
