+++ sources/ressources/textures/textures/texture_init.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   textures.c                                         :+:      :+:    :+:   */
+/*   texture_init.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/23 13:53:13 by olimarti          #+#    #+#             */
-/*   Updated: 2024/02/11 07:50:13 by olimarti         ###   ########.fr       */
+/*   Created: 2024/02/19 22:23:11 by olimarti          #+#    #+#             */
+/*   Updated: 2024/02/19 22:23:45 by olimarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,28 +14,6 @@
 #include "maths_utils.h"
 #include "structures.h"
 #include <assert.h>
-
-void	texture_update(t_texture *texture, size_t dt)
-{
-	texture->current_time += dt;
-	if (texture->current_time >= texture->ms_per_frame)
-	{
-		texture->current_frame = (texture->current_frame + 1)
-			% texture->frame_count;
-		texture->current_time = texture->current_time % texture->ms_per_frame;
-	}
-}
-
-t_img_data	*texture_get_frame_offset(t_texture *texture, int offset)
-{
-	return (texture->frames[(texture->current_frame + offset)
-			% texture->frame_count]);
-}
-
-t_img_data	*texture_get_frame(t_texture *texture)
-{
-	return (texture->frames[texture->current_frame]);
-}
 
 int	texture_init(t_texture *texture, int frames_count, int ms_per_frame)
 {
@@ -47,19 +25,4 @@ int	texture_init(t_texture *texture, int frames_count, int ms_per_frame)
 	texture->ms_per_frame = ms_per_frame;
 	texture->current_frame = 0;
 	return (0);
-}
-
-void	texture_destroy(t_texture *texture)
-{
-	if (texture)
-	{
-		free(texture->frames);
-		texture->frames = NULL;
-	}
-}
-
-void	texture_set(t_texture *texture, int frame, t_img_data *img)
-{
-	assert(frame < texture->frame_count);
-	texture->frames[frame] = img;
 }
