+++ sources/ressources/textures/textures/textures_set.c
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   textures_set.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/23 13:53:13 by olimarti          #+#    #+#             */
+/*   Updated: 2024/02/19 22:25:14 by olimarti         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "draw_utils.h"
+#include "maths_utils.h"
+#include "structures.h"
+#include <assert.h>
+
+void	texture_set(t_texture *texture, int frame, t_img_data *img)
+{
+	assert(frame < texture->frame_count);
+	texture->frames[frame] = img;
+}
